采集器可靠性与隔离

+++ config.h
@@ -6,8 +6,8 @@
     double interval_sec=2.0;
     double cpu_threshold=80.0;
     double mem_threshold=80.0;
-    double alert_consecutive_count=3.0;
-    double alert_cooldown_sec=60.0;
+    int alert_consecutive_count=3.0;
+    int alert_cooldown_sec=60.0;
 };
 
 SysmonConfig loadConfig(const std::string& path);
